filterx/jsmn.h: add a header-only JSON parser
Jsmn is a simple JSON parser that does not have a object layer, rather it
allows the user (e.g.  syslog-ng) to use its preferred representation to
store JSON objects.  This maps more nicely to filterx dicts than the
previous implementation, where we loaded the JSON first and then mapped the
in-memory data structures to filterx.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>