logmsg/nvtable: improve nv_table_get_next_size() for incremental growth
Whenever NVTable needs to grow (either during clone or realloc), the new
nv_table_get_next_size() function calculates the right new size, while
ensuring some "additional_space" becomes available in the new clone.

Previously, we always erred on the side of increasing the allocation size,
so that any expected new changes would not need another bump of the size.
This algorithm however, can waste a lot of memory.

The new algorithm is:
  * increase at least 256 bytes,
  * only increase the size if the current allocation does not fit
  * use 4k aligned sizes for large NVTables (>4kB) or 1k for small ones

We might have slightly more reallocs this way, but the total allocation of a
LogMessage should be a lot less.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>