tests/test_thead_wakeup: remove this unit test
The test is flaky, and is not really testing anything specific. It is coming
from the time when syslog-ng was converted to be multi-threaded and
probably there's an assumption that accept/read would be interrupted by
the delivery of a signal. This is a fair assumption about most
modern platforms and I am not even sure where we assume this in the
codebase. So even if this fails in the future, we can't easily pinpoint
anything to fix.

It's better removed.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>