filterx/filterx-ref: fix crash caused by false sharing of mutable objects
In case a dict or a list is shared, we can't use the FilterXRef instances
stored in the shared dict, even if they are part of the same recursive
structure.

If we did, we would end up with an infinite data structure, that causes a
crash due to stack overflow when iterating the data structure.

Reproduction:
	filterx {
            d = {
              'sub': {},
            };
            d.foo = 'fooval';
            d.sub.recursive = d;

	    $MSG = string(d);
	};

The above crashes when trying to turn "d" into a JSON and iterating the dict.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>